Integrated the Order API Controller
I've added a new DTO to the Web API project because the API call parameters are limited to one.
It derives from OrderDTO and simply declares an ICollection of OrderProductDTOs.
I was considering extending the existing OrderDTO in our Datalayer, however, it's a small variation and is only needed in the API so, to me, it wouldn
t make sense to extend. That's why I made the change.

- Added
  - OrderController - All Tests successful
  - OrderWithProductsDTO
- Fixed
  - BrandController and CategoryController not inheriting from ControllerBase