Layout for Product info and Navigation to it.
I've made the layout for the `Product info` in a componenet named `ProductDetails`.  You can navigate to this page, by clicking on the `Title` of the `Product` out on the `Index` page. That is the purpose of it but, this is still to be tested as the `OnInitializedAsync()` isn't functioning properly.

 - Added:
   - `ProductDetails` page which will work as a view of a single `Product` and with the `Image` being a bit larger than on the `Index` page.
   - To the `ProductCard` i've added a button to the `Product.Title` so that a `Customer` can navigate over to the `ProductDetails` page, with the `ProductID` following.

OBS. we need to implement the functionality to the `OnInitializedAsync()` in the `ProductDetails` page, for it to even display anything - as this will be showing a specific `Product`.