Layout for Edit page and setup for functionality.
I've implemented the `EditProduct` page with the layout, so that it will have a form with a submit. The `OnInitializedAsync()` is still not functional but, I've set it up to work as intended when updating an object.

I've made sure that the page will have a routing value of `ProductID` and we're retrieving that in the `OnInitializedAsync()`. However, this ain't tested properly as it won't load the Web API properly.

 - Added
   - `EditProduct` page with matching layout, this time including an `EditContext` to the `EditForm` as we're making a `Put` request, rahter than a `Post`. I've also set that the inputs will be the new values in the `product` that we've retrieved. Lastly ending with sending the `ProductForm` through our `EditContext`.