Added all base containers for pure API calls
I've added all the base containers to mach the pure form of JSON recieved from /Controller/[Object]/GetBy/{_[Object]ID}.
This will make it much easier for us when using the API.
We can then use interfaces to mask properties or extend the base.
One example included with the changes is IProductDisplay, which extends the ProductBase to include associated images as URLs.
- Added
  - BrandBase
  - CategoryBase
  - CustomerBase
  - OrderBase
  - ProductBase
  - IProductDisplay
  - ProductDisplay: Derived from ProductBase and serves as a displayable product where all associated images are attached