CRUD functionality to `Product` has been added.
I've added a `Create` button on the index page, where you can create a new `Product` and it will be shown on the index page when it has been created.

You can access the Detailed view of a `Product` by clicking on the `Title` in the index page, which will show the information of the given `Product` there has been chosen, with additional buttons. a `Add to basket` button, which has no functionality yet and a `Edit Product` button, which will redirect you to the Edit page for that given `Product` .

 - Added:
   - `BrandBase` and `IBrandBase` Model and interface. These are used when the `product` are being displayed in the `ProductDetails` page as these aren't shown out on the index page.
   - `CategoryBase` and `ICategoryBase` has been implemented for the same purpose as above.
   - `ProductCreate` and `IProductCreate` has been implemented to use over on the `CreateProduct` page for the `EditForm` that is going to be pushed up to the DB. This is to keep our Models unique to óne function.
   - `EditableProduct` and `IEditableProduct` has been implemented for the same reason as the above but, for out `EditProduct` page instead.
 - Modified
   - The index page now have a button with the `CreateProduct` functionality, which will redirect the user over to another page, for them to create a new `Product`.
   - Changed the `Product.Title` to be a button which will redirect the user to the `ProductDetails` page about that single `Product` with a method `GoToProduct()`.